+++ src/PanelTraits/Read.php
@@ -185,6 +185,25 @@ trait Read
     }
 
     /**
+     * If a custom page length was specified as default, make sure it
+     * also show up in the page length menu.
+     */
+    public function addCustomPageLengthToPageLengthMenu()
+    {
+        // If the default Page Length isn't in the menu's values, Add it the beginnin and resort all to show a croissant list.
+        // assume both array are the same lenght.
+        if (! in_array($this->getDefaultPageLength(), $this->page_length_menu[0])) {
+            // Loop through 2 arrays of prop. page_length_menu
+            foreach ($this->page_length_menu as $key => &$page_length_choices) {
+                // This is a condition that should be always true.
+                if (is_array($page_length_choices)) {
+                    array_unshift($page_length_choices, $this->getDefaultPageLength());
+                }
+            }
+        }
+    }
+
+    /**
      * Specify array of available page lengths on the list view.
      *
      * @param array $menu 1d array of page length values,
@@ -217,6 +236,8 @@ trait Read
                 }
             }
         }
+
+        $this->addCustomPageLengthToPageLengthMenu();
 
         return $this->page_length_menu;
     }